Intern_Submissions/Prajapat_kavitha/requirements.txt
torch
transformers
accelerate
sentence-transformers
langgraph
faiss-cpu
pypdf
numpy
duckduckgo_search
gradio
markdown-pdf
